+++ packages/plugin-truffle/src/utils/options.ts
@@ -19,7 +19,7 @@ export interface DeployOptions extends Options {
 }
 
 export interface UpgradeOptions extends Options {
-  call?: { fn: string; args?: any[] } | string;
+  call?: { fn: string; args?: unknown[] } | string;
 }
 
 const defaultDeployer: Deployer = {